backend/requirements.txt
# FastAPI y servidor ASGI
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Base de datos y ORM
sqlmodel==0.0.22
asyncpg==0.30.0
alembic==1.14.0

# Autenticación y seguridad
pyjwt[crypto]==2.10.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

# Utilidades
python-dotenv==1.0.1
pydantic[email]==2.10.6
pydantic-settings==2.7.1

# Desarrollo (opcional)
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
